+++ src/const.js
@@ -23,9 +23,24 @@ const SortType = {
   TIME: 'time'
 };
 
+const NoPointsTextType = {
+  [FILTER_TYPE.everything]: 'Click + New event to create your first point',
+  [FILTER_TYPE.future]: 'There are no future events now',
+  [FILTER_TYPE.past]: 'There are no past events now'
+};
+
+const Method = {
+  GET: 'GET',
+  PUT: 'PUT',
+  POST: 'POST',
+  DELETE: 'DELETE'
+};
+
 export {
   UserAction,
   UpdateType,
   FILTER_TYPE,
-  SortType
+  SortType,
+  NoPointsTextType,
+  Method
 };
